Add a 'safe mode' to Sling.
When safe mode is enabled, all Sling methods copy the underlying
sling instance before mutating the Sling object effectively making
Sling immutable.

This is just a proof of concept to see whether it makes sense.  It needs more
documentation and testing before it's shippable, but I wanted to get feedback
before doing that work.

I think ultimately this mode with be inverted as unsafe mode so that we do the
least surprising thing by default and allow clients to opt-in to a mutable
world, but I've erred on the side of not changing existing behavior.