Add response data streaming facility
Here we add a new ResponseDecoder: ByteStreamer. This allows user to get
a stream of bytes 'as is' i.e. without any decoding. This is useful for
all those streaming use cases like downloading files, binary blobs etc.

This is also useful when response data is just plain text for e.g. the
5XX responses from API gateways.

Side changes:
- The xml stuff in testing code no longer needed as this new
decoder is used as non-default decoder.

- Code reformatting as per gopls/gofumpt style.